Implement removing geofences, timer, and reboot receiver
Geofences will be removed after they're triggered by sending
a broadcast to HospitalTrackingService from the receiving
IntentService. A 24-hour timer is setup after geofences
are created to refresh the geofences. This is to account for
any drastic location changes (i.e. user flew to a different
city overnight), and to reintroduce any triggered geofences that
were removed. The reboot receiver is triggered when the system
reboots. It restarts the tracking service because the geofences
are cleared when the system shuts down.