Refactor service logic
Instead of having an eternal service, shut down the service after
performing its task. IntentService was not an appropriate option
for this because there are many callbacks that need to be waited on.

Additionally, an Alarm is set when the user first opens the app
or when the device reboots. When the alarm triggers, it will call
the service to re-create geofences.