Refactor hospitalization logic
Users must DWELL in a hospital for 4 hours to qualify as being
hospitalized. If the 4 hour requirement is met, the user will
be presented with a survey 1-hour after leaving the hospital.

All data stored in SharedPreferences is now encrypted.

A daily alarm will be scheduled to make sure the registered
geofences are within the user's location. If the user has
traveled a significant distance, we'll retrieve a new set
of hospitals.